VideoPlayer: iOS, rename RendererVTB to RendererVTBGLES
