+++ xbmc/cores/VideoPlayer/VideoRenderers/HwDecRender/RendererVTBGLES.h
